requirements.txt
Pillow==9.5.0
qrcode==7.4.2
Requests==2.30.0
